+++ api/channels.go
@@ -158,7 +158,7 @@ func (c *RestClient) DeleteChannel(ctx context.Context, channelID discord.Snowfl
 }
 
 // GetChannelInvites returns all invites for a channel. Requires MANAGE_CHANNELS.
-func (c *RestClient) GetChannelInvites(ctx context.Context, channelID discord.Snowflake) (*[]*discord.Invite, error) {
+func (c *RestClient) GetChannelInvites(ctx context.Context, channelID discord.Snowflake) ([]*discord.Invite, error) {
 	if err := channelID.Validate(); err != nil {
 		return nil, err
 	}
@@ -168,7 +168,7 @@ func (c *RestClient) GetChannelInvites(ctx context.Context, channelID discord.Sn
 		return nil, err
 	}
 
-	return doRequest[[]*discord.Invite](c, req, map[int]bool{
+	return doRequestSlice[discord.Invite](c, req, map[int]bool{
 		http.StatusOK: true,
 	})
 }